[aptitude @ Fix message translation function P_()]
Sometimes translatable messages contain an additional classification string
useful for translators to analyse the usage of the string. Example:
P_("Encoding of README|UTF-8")

The part after "|" is the translatable string, the remaining string is the
explanation. A translation of the part after the pipe character only was
ignored. It was also handled wrong if the Native Language Support was disabled.