Implement radio configuration buttons in the new list-style configuration editor.
After some thought, I decided to go with several siblings to implement
a single radio item, instead of a single object showing the current setting
that can be modified with left/right.  IMO, this approach is more friendly
since all the options are visible up-front (although you have to select
individual choices to see the long description of each choice).