Add a default rejection of solutions that would remove Essential packages. (Closes: #486748) This is slightly unideal in two ways that should be fixed eventually.
First, the UI is less than desirable if this leads to a situation where
nothing can be resolved: the user just gets the message that
dependencies can't be resolved, with no indication why; perhaps there
should be some explanation that some versions were rejected.  Second,
the removal of Essential packages will appear in the UI as an alternative
resolution, just one that's rejected by default; I'm not sure if this is
good or not.  Third, the UI should probably continue to do a last-minute
check for Essential packages before running an installation; I'm not
fixing this right now because it's more complicated to do so (to do it
right you should remember which Essential packages the user said it was
OK to remove and only remove those, and maybe also remove them from that
list if the user reinstates them).