Don't show the "version" column in package lists or preview views. Experimental UI tweak.
I want to try this because the (a) versions take quite a bit of screen
real estate that could be used for other things, such as the package
short description, and (b) versions are just extra visual clutter when
the user is trying to find a package.

The version is interesting information and so it should still be
displayed in the "info" pane, but it's not important enough to warrant
consuming scarce horizontal space in the list for it.  The package list
should (by default) contain just the most important information about
a package: an overview of the package's state, information that's
useful for comparing packages, and information that helps the user
figure out whether each package is relevant.  While the version of a
package is "important" in some sense, the version string by itself is
not generally a useful thing to know and it's particularly useless for
cross-package comparisons.  If the user iis interested in a particular
package's version, she can just click it in the package list.