Make the preview notification global, show how many packages are broken in it, and let the user jump to the resolver using it. Perhaps we should also support showing which packages are broken?
This patch also tweaks the notification API.  That should probably have
been in a separate commit, but I was too far into it before I realized
that.  The notification abstraction is now a bit tighter: the protected
members are private in accordance with the general practice in aptitude,
and I removed the get_buffer() method for the time being.  (we can add
it in later if necessary, but I think that most users of the Notification
class will mainly want to change all the text at once)