Redesign how matches are represented to handle version information cleanly. I think this is a reasonable final version of the match structures. It uses a two-level tree structure, where the outer level represents the parts of the pattern that don't match particular versions of particular packages, and the inner level describes the actual versions that were matched and which patterns matched them.
Still to do: the code to actually generate these things needs to be
written; the main open question here is how completely I'll fill in the
match structure.  I might try some sort of three-level match if I can
find an efficient implementation: at the lowest level, it just returns
a bool ("yes it matched" vs "no it didn't"); at the next level, it
returns some match information but with short-circuiting (so some
versions that could have matched will be left out), and at the highest
level it will find every version that would have matched; e.g., all the
clauses of an "or" will always be examined, and in "any" mode all the
matchables in the current pool will be examined (even if a match is
found).

I still wonder what to do about the children of a "not", but for now I
just won't describe them.