Massive refactoring and redesign to properly support background downloads in the GUI and running dpkg in a terminal. This should probably have been split into a lot of smaller commits, but it's too late now.
The GUI now runs downloads in a background thread, just like the curses
interface.  In fact, it is *just* like the curses interface: the code of
ui_download_manager and download_thread has been parameterized so that it
can be used from either the curses frontend or the GUI frontend.  These
files should eventually migrate to the generic directory.  Some of the
parameterization left ugly stuff behind, like refcounted_progress and the
wacky redesign of the curses progress class (which took place solely so
that it could be unified with the GTK+ one).  It may be worth taking some
time to clean this code up.

In addition, the code for download_install_manager has been refactored
again.  Instead of exposing the different parts of "finish", which would
eliminate any hope of having a unified ui_download_manager harness, the
"finish" method (for all download managers) is now explicitly continuation
based, and the install manager uses this feature to break up the download.
The function that runs dpkg in a terminal now takes a continuation to be
invoked in the main thread, so it can fork dpkg in a subprocess, and then
call the continuation when the subprocess terminates.  This means we don't
need to mess around with threads at all in the dpkg-terminal-creation
process, a huge relief.  It also fits in with the generally continuation-y
feel of a lot of the aptitude code, and everyone loves continuations in
their GUI programs anyway!

Note: in this commit, I've adopted a policy of always returning right
after I invoke the continuation, to make it clear in the code what's
going on and to minimize the chances of a future edit screwing this up.
(if I really wanted to get serious about that, of course, I would write
an INVOKE_CONTINUATION macro...)

A few other minor things were touched up here: for instance, the program
no longer enters an infinite loop after running dpkg, and more download
events are logged.  There's still more work to do around the edges:
changelog downloading from the GUI is broken, and the download status
tab is all messed up.  Also, I'd like to support the status pipe from
dpkg (but I think that's a post-0.5 thing).