Add a "promotion set" object. This is the first step towards implementing the tiering system laid out in <20090214163554.GA8343@emurlahn.burrows.local>.  "promotions" are the evolution of "conflicts" in the current system: they store the knowledge that "if you do this and that and that, you must push the solution to tier N".  The "promotion set" will take on the role of the current "conflict set"; because the behavior we need is quite a bit more complex, I've chosen to write a custom data structure rather than leaning on the dense_setset template.  That means more work (it took me all day just to implement this initial version) and more debugging will be needed, but on the other hand I've already stuck in some simple optimizations that were obvious once I was working with promotion elements and not just generic "value objects".
Unlike the existing conflict set, elements can be removed from the
promotion set.  In fact, if you add a new promotion that supersedes
an existing promotion, the existing one will automatically be thrown
out.  You can also remove all the promotions on one tier, or all the
promotions below tier N.  The only reason for not having more general
erasure is that it's not needed, and by providing the specific cases
that I need directly, I can optimize them a little better.

The unit test verifies that a promotion set can be created and that it
sort of works, but it needs more work before I can have any real
confidence in this code.  Luckily, the currect code also isn't used
anywhere.  Yay!