First step in the actual implementation of tiering. This commit introduces solution tiers, replacing the old conflict, deferred solution, and already-generated-solution structures with a system based on promotion tiers.  Currently this isn't used to actually order resolver results; however, it is used to accumulate higher-order information about deferals that couldn't be calculated before, and to accumulate some information about soft dependency breaks.  (currently they are still somewhat restricted, though)
The next step is to push the use of the new choice structure outward,
and maybe to profile the new code: it achieved a four-fold reduction
in the number of search nodes visited in a moderately large upgrade
test, but doesn't perform any faster in real time.  I'm guessing
there are some inefficiencies there that I need to look at.