Restructure the resolver so that it stores and returns sets of choices using choice_sets, rather than manually storing the different types of choices in the solution object. This has two beneficial effects.  First, it should make it a bit easier to add a new choice type, although many parts of the program will still have to be touched.  Perhaps more importantly, this provides a cleaner conceptual model of what the resolver does: it finds "choice points" (that is to say, dependencies that need to be solved) and makes choices at them.
This commit sneaks in a few optimizations that I came across in the
process of rewiring all the resolver logic.  The big one is that testing
whether a solution violates a user constraint now requires steps
proportional to the size of the solution, not to the number of user
constraints.  (each step is logarithmic in the number of constraints,
but that's still a big improvement)