Write up a draft design for the "jump" concept, which will be used to model package replacement. In practice, jumps will exist from the removal version of a package that has replacements to every version that replaces that package.
Jumps have some similarities to soft dependencies, and they could be
represented by them, but there are important semantic differences.  The
biggest one is that a dependency is a constraint and a jump is not.
Instead, a jump is a rule that tells the solver how to structure its
search: whenever it places a particular version into the solution, it
should split the search to consider several other versions as well.
Among other things, this distinction means that the resolver will not
attempt to spontaneously install replacements for packages that aren't
installed; it will only add replacements to the solution when the
solution also contains the removal version.