Make tiers be tuples of integers (lexicographically ordered) instead of plain integers. This allows us to classify solutions in a more fine-grained way.  In particular, it provides a not-horrible way to integrate apt pin priorities into the resolver's decisions, by making them a subordinate level of ranking to the main tiering system.
The current plan is that the first level of tiering is the only one the
user sees (by default) and the only one that can be directly configured.
The subordinate levels are filled in automatically (pin priorities are
all I have in mind now, but I don't discount the possibility of
introducing more levels in the future).