Add support for "dependency-scoped" install-version choices, which match the same install as long as it was done for the same dependency, and use them to inject promotions related to approval user constraints. Approvals are violated by search steps that choose a different solution to a dependency than the approved one.  So to represent these in the condition of a promotion, we have to be able to match installing a particular version *but only if the install was for a particular dep*.
In other words, install_version choices now have four levels of
specificity:

  (no dep) -> (unscoped) -> (dep-scoped) -> (from-dep-source)

Choices with no dep "contain" any other choice that installs the same
version, and so do choices that are unscoped (but have a dep).  Choices
that are dep-scoped "contain" choices that have the same dep (so no-dep
choices are out, but some unscoped choices can match).  Choices in the
from-dep-source category "contain" choices that have the same dep and
that are also from-dep-source.

This commit adds the code necessary to represent these extra types
of choices, to manipulate them in the resolver, and to parse them in
the log analyzer.  It also uses these new abilities to generate
promotions to the "defer" tier when successors violate a user
constraint.