Regenerate approval information when a new mandated version or forced-broken soft dependency is added. Unfortunately, this is pretty obviously necessary for correct behavior. Approving a version can cause another version to become legal (it's OK to "avoid" an approved version by installing another approved version), so we have to figure out which of the previously deferred solutions are "really" deferred.
This would be a good excuse to write that code that tracks the reason
for each deferral and strips out only the information that might have
to go.  Most of the time there should not be a need to throw out
deferral information, but we don't have any way of knowing this right
now.