Overhaul the resolver tab. Eliminated the previous/next holdover from the curses interface: it made sense there because of the limited real estate and input mechanisms, but in the GTK+ frontend it's much better to have a list of the already-generated solutions and a button to make a new solution.
There are still a bunch of loose ends.  The display of the current
solution doesn't adapt well to being narrow -- I think the "action"
column should be nuked and merged with "name", but I'm not sure what
else to do there.  Display of actions is somewhat inconsistent.  It
would be nice to unify them, but we should do that with an eye to the
fact that differing renders are also useful.  At the very least, use
of text styles should be straightened out: I'm not sure that showing
version numbers in a larger font works, but we should either do it
everywhere or nowhere.  It's scattershoot now partly because I wanted
to throw lots of stuff at the wall and see what stuck.  I also would
not bet much money on the behavior being absolutely correct.