Abandon reinstatement of unused packages. (Closes: #522881, #524667) The basic problem here was: A conflicts with B, which is automatically installed on the system and held by C.  An upgrade installs A and removes C (removing B as unused), but dependency resolution decides to put C back on (but not B -- maybe it's only recommended or there's an OR involved).  Now B will be reinstated, but it conflicts with C => dependency error.
This means that sometimes packages will be "removed because they are
unused" when in reality they are still used because a dependency
solution put their dependers back onto the system.  But IMO that's
better than dependency solutions that produce a broken system.  Ideally
the dependency solver would know about autoremoval, but that's a rather
tricky nut to crack open.

The worst negative effects of this could be mitigated by calculating
a solution before autoremoval happens (in the command-line interface).
That could be done either by using hypothetical reasoning (like the GUI
upgrader does) or by just using an action group to defer removal --
probably the latter is easier.  In either case this isn't a perfect
solution, since it could lead to spurious dependency errors (stuff that
could just be dropped, but conflicts with newly installed packages) and
it could lead to too *much* stuff on the system (e.g., something is
installed for a dependency, the depending package is unused and gets
dropped, but a virtual dependency holds the dependee on the system :( ).