Kill off dep-scoped choices. These were always a little weird, they turn out not to be necessary for the purpose they were originally introduced to serve, and they play havoc with some of the optimizations I'm working on.
They aren't necessary to generate promotions for approvals, because
only the deferred choice itself would get this special tag.  Just by
the usual way promotions are generalized, the dep-scoped choice is
always eliminated before it does anything useful.

The specific problem motivating this change is that dep-scoped choices
make it much trickier to have a mapping from choices to lists of solved
dependencies using the generic choice_indexed_map object.  This isn't
an insurmountable problem, but considering the other issues with
dep-scoped choices, I thought it was a good idea to get rid of them.
For instance, look at how much simpler choice_indexed_map is when it
doesn't have to support dep-scoping.