When applying an incipient promotion, use the promotions's choice rather than any choice from a solver. This should be what we do anyway: the promotion's choice will always be the most general one that covers all the solvers that hit it.
However, the fact that the previous code wasn't working exposes a bug in
contains(): although deps are no longer significant, it's claiming that
one choice with a dep doesn't contain another choice with a different
dep, even if neither is a from-dep-source choice and they install the
same version!  That will be fixed momentarily.