Only generate entries in the global reverse index of steps when a choice is first added. The global index was getting ridiculously huge, and it doesn't need to be.  To find all the steps containing a choice, we can just start at the first step where it was introduced as a solver, and trace out the whole subtree below that, stopping whenever a step no longer contains the choice in question.
With this commit, the time per step is much more reasonable (when
compiled without optimizations, it appears to be approximately as fast
as the old resolver was with optimizations).  However, it also looks
like there are a few bugs left: it branches far too much and it got
stuck once in an infinite loop.