Ensure that the first time we add a successor to a step, we do not update the "is_last_child" member of the last step in the steps list. This may be a symptom of the fact that the way steps are added is rather error-prone.  However, it should work now.
The problem was that we were always clearing the "is_last_child" member
of the last step in the global list of steps.  We want to do this
whenever we add a child beyond the first, since otherwise each step
would have only one child.  But we don't want to do it for the first
child, or only the very last step in the list would be marked as a last
child (meaning that every "child" list would contain all the steps
following its first entry).