Don't attempt to track tier validity at the solver level; instead, when a promotion is retracted, recompute tiers for solvers that it might have "hit". In this model, the tier_valid member of search_graph is no longer an active expression that recomputes the solver's tier when it becomes invalid.  Instead, it's a pure expression that represents whether the last promotion that was applied to the tier was valid.  When any promotion becomes invalid, the whole list of active steps is searched for steps that it might have influenced (the same way we would search for steps to apply it to) and the tier of each such step is recomputed.
The major reason for making this change is that maintaining all the
structures to track which solvers a promotion affected was expensive,
particularly in terms of memory usage.  This information is only needed
rarely, so I don't want to overly burden the normal behavior of the
resolver with book-keeping to handle the corner case of retracted
promotions.  The new code should be correct; it'll be a bit slow to
retract promotions and sometimes the wrong solvers will be recomputed,
but that's better than eating hundreds of megabytes of memory.

A secondary, but also important, reason is that this opens the door to
memoizing dependency solver sets.  Previously, solver sets contained
backpointers to their own step; they had to, in order to maintain the
"zap this solver in this step" active expression.  Now, solver sets
only contain pointers to structures that are common to all solvers for
their dependency, or to all solvers that were affected by a particular
promotion.  This should mean that there are enough equivalent sets that
memoization of some sort will be a win.  In particular, I hope that this
will tackle the last outstanding memory problem, where applying a
promotion takes a bunch of identical sets that share memory and creates
a bunch of identical sets that don't share memory, massively spiking
the program's memory usage in the process.