Use pointers to package cache structures to store resolver objects, rather than iterators. This cuts the runtime memory use down by 20-30%.  The problem was that a *huge* number of these objects are allocated, and iterators store members that either aren't needed, or are redundant with other members of the containing resolver object.
The dep-cache pointers could, perhaps, be dropped, if I forced
the routines that want one to request a "universe" pointer instead.
On the one hand, that's just about everything now; on the other hand,
that would chop many of these structures down to a single pointer!
Worth a thought, anyway.