Fix recompute_step_tier() to correctly compute the tier. It wasn't taking into account the actions themselves, or promotions that only included them.
This fixes the big glaring problem with the new code, namely that it
was spinning off indefinitely (because depending on the order in which
some normally idempotent things were applied, they would trip this bug
and miss the fact that a line of reasoning was supposed to be chopped
off).  There's still one odd problem, where I get different results
each time I run it, probably because of some other order dependency.