When increasing the tier of a solver that's newly deferred, only increase exact matches (including matches to the dependency). Necessary because deferrals due to violating an approval constraint only apply when a version is chosen to solve the same dependency.
This is just one other way that approval constraints are wonky.  I'm
thinking that it might be good on multiple levels to throw them out:
they're a bit counterintuitive, they tend to produce unstable results
(the set of legal successors can vary depending on which dependency
is chosen as the next one to resolve, for instance), and because they
mean that deferrals happen on a per-dependency basis, they need a lot
of careful code and a lot of extra memory allocation to get right.

I'm not sure what would replace them, though.