Write untested implementations of getItem and putItem in the file cache. In the process, I tweaked the schema and supporting code as things came up.  The most notable change is that I now just use the auto-incrementing primary key field to check which cache entries are least recently used, rather than relying on the date (this makes it easier to drop a bunch of old entries at once).
The new code probably needs to be factored better -- getItem() and
putItem() are long and have a lot of repetitive error handling.  Maybe
the error handling should be done with a wrapper routine? (i.e., "run
this thunk in a transaction; if it throws an exception, roll back")

Also, I need to handle erroneous situations better.  It's not a fatal
error if something goes wrong in the cache -- it should just mean that
something can't be cached or retrieved.  If the database is messed up
in a way that causes this to happen, maybe we should just delete it and
create it from scratch instead of failing over and over.