Ditch increase-tier operations due to poor interaction with add-to-tier. The comment at the top of tier_operation cheerily asserted that these two operations formed a specific partial order.  They don't.  Each of them *individually* does, but there is no natural ordering between them: for instance, "increase to 100" takes 5 to 100 and 105 to 105; "add 5" takes 5 to 10 and 105 to 110.  Worse, they aren't closed under composition, since order matters and it isn't stored in an operation (and can't be without making operations contain a whole chain of compositions).
The effect I was trying to get with increase-tier can be achieved,
albeit not in a backwards-compatible way, by adding levelwise upper
bounds to tier objects at a higher level (i.e., in the resolver).
The point here is to be able to have a tier that, for instance, is
0 if there are no removed packages and 1 if there are *any* removed
packages (i.e., it doesn't distinguish between different numbers of
removals in the same way that the current removal tier doesn't).