Introduce new hint types to tweak cost components. This also starts to introduce the concept of a "cost component": an intermediate counter of some sort that can be used to assemble a cost. The eventual goal is to haved named components, like "removals" and "upgrade-cancels", and operators that combine them to form new cost components, like "removals + upgrade-cancels".
"cost component" / "cost" is the terminology I eventually want to move
towards, in favor of the confusing and badly-chosen "tier level" and
"tier".  This commit starts to introduce that terminology, and all my
new commits will use it (except where local consistency demands
otherwise).  Eventually I need to make a pass over the code and rename
lots of stuff.