Add a suite of parser combinators. Still needs more tests.
This is heavily inspired by ghc's Parsec library.  In particular,
it differs from Spirit in that the "value" of a parser is a core part
of its definition, and in that it's very reasonable to implement your
own parsers by hand (I didn't explicitly say this in the doccomments --
maybe I should -- but the sequencing operator >> is equivalent to
invoking the parsers in successive statements, like in a Haskell
monadic parser.

Probably not as efficient as Spirit, but should be a lot easier to
use (and the places I want to use it only parse short strings anyway;
even a fairly slow library would probably be OK).