Fix a potential performance problem: the cost of solver sets was being underestimated because the "maximum" cost isn't actually maximal. The algorithm that was at fault started with the highest possible cost and incrementally lower-bounded it with the costs in the solver set, to compute the lower-bound of the whole solver set.
That was the idea, anyway.  It probably worked right when tiers had
exactly two elements and there was a sensibly-defined highest tier.
But now that means we start with

(MAX, 0, 0, ...)

and then lower-bound with, say,

(10, 1, 5, 4)

to get

(10, 0, 0, ...)

and, well, you see how that goes.  The fix is a bit of a hack but will
do what I need here (producing a conflict if there are no solvers and
the lower-bound of all the solvers' costs otherwise).