When preparing a variant build environment, allow flag values to contain environment variable references. This involves two things: flags have to be set up after environment variables, and we need to ask the environment to substitute for variables when merging the new flag values.
The main reason for doing this is that it makes it easier to have a
single configure step setting up compile flags that several build
variants share.  We can put the flags into a variable, then merge
that variable into the compile flags only in the variants where they're
relevant.

For instance, aptitude will have a GTKFLAGS variable containing flags
for GTK+ compilation; the GTK+ variant will add $GTKFLAGS to the
environment.