Redesign the dynamic-list interface to be more general, and update its tests to match. Specifically, I changed append() to a general insert() routine and added a move() routine, updating signals to correspond.
I decided that although this imposes a bit more on clients of the
interface (they need to support more signals and more general signals),
it's probably better in the long run.  It makes the list interface less
surprising; it also makes it possible to more cleanly assemble several
lists into a larger list (something that was ugly with the previous
interface).

Enumerators were eliminated because I forsee a need for direct indexing,
and having both mechanisms was just clutter in the interface (IMO).

The list collection object is now broken, and its tests are temporarily
commented out.  I also need to write a test for the new move() method.