Overhaul the area code to prepare for actually using it. Three changes here.  The smaller one is that it now uses dynamic_set objects instead of dynamic_list objects to store tabs and notifications. The medium one: explicit signal objects have been replaced with interface methods for connecting and/or invoking the signal; most interfaces hide one or the other of these functions.
  The large one: tab_info has been redesigned and split into two "role"
interfaces: one showing what the view displaying the tabs should need,
and the other showing what the tab itself needs (to update its icon,
progress, etc).  This makes it much clearer what the intended
relationships among the various components of this subsystem are, and
makes it clear in a way that the compiler can enforce.  Sadly, this
means I need to use multiple inheritance. (note: it would be less tricky
if I didn't use sigc::trackable -- maybe avoiding that is a good idea?)