Add a first draft of a logging library to replace liblog4cxx. Currently unused.  Needs test code.
The reason for replacing log4cxx is that on the one hand, it
provides much more functionality than aptitude needs in any
conceivable universe, and on the other hand, it seems to be
buggy and fragile.  Just initializing it takes a fairly careful
dance, and it seems to be behind at least some of the sporadic
crashes users are seeing in aptitude.  This new code does
exactly what aptitude needs, in a simple and hopefully much
more robust way (possibly a bit slower -- but the fast path
should still be fast; any slowdown *should* be limited to
the situation where we're logging a ton of data).