dd some generic code to display a progress indicator on a single line. This is just a slightly higher-level wrapper around the transient message class.
Preliminary work with this suggests that the progress_info API might need
a redesign; I see a significant slowdown when using this to report progress.
I don't think this is insurmountable; I just need to be sure that I actually
handle it before I push the code that uses this into "master".