Add a first draft of some wrapper code to display search progress events. The design needs to be improved, but it's a proof-of-concept.
The wrapper takes care of displaying which pattern is being searched for
as part of the displayed progress message.  This has the side effect of
causing the construction of progress messages to be a substantial fraction
of the overall run-time.  The API needs to be redesigned so that messages
are only formatted when they'll actually be displayed.