Make progress_info objects mutable. Currently the type can't be changed, because that could introduce weirdness into the API and it's not really necessary.
The reason for making them mutable is that it's common to want to update
the current progress fraction in a tight loop.  By using mutable objects,
we can avoid having to constantly copy the status string.  I clocked nearly
a 10% speed increase with this change, and it makes the overhead from the
progress display negligible in my tests.