Replace the db2latex generation for the docs with fo -> pdf generation.
db2latex is pretty much dead and built on a deprecated stack.  I
still wish there was a way to get decent latex output from docbook,
but it seems that we're in a "never the twain shall meet" situation
for now.

Note that aptitude's documentation can't be processed through
the packaged version of fop until a bug in fop is fixed (filed
in debbugs with a patch).