Update package lists using ListUpdate (libapt-pkg)
The previous list update code is mainly a duplication
of the algorithm available in libapt-pkg.  It has a
history of being problematic and was generally never
up-to-date with APT.  The new code simply wraps an
algorithm provided by libapt-pkg and resolves some
long-standing issues:

- Update errors are reported. (Closes: #451137)

- Run APT::Update hooks. (Closes: #476399)

The one known short-coming with the new code at the
moment is that if the user cancels the download in
the curses interface, errors are reported for any
files that were remaining.