Update/fix po4a.cfg.
Most importantly, apply the addenda to aptitude.xml instead of to
manpage.xml, since the authorship information which used to be
contained in the latter is now inherited from the former, and this is
what the addenda are intended to be added to.

Also: use [po_directory] and $lang to remove most of the repetition,
update the "opt" string to avoid warnings from po4a, and use it for
aptitude.xml as well as for manpage.xml.